+++ cpp/src/arrow/dataset/scanner_test.cc
@@ -34,6 +34,7 @@
 #include "arrow/testing/future_util.h"
 #include "arrow/testing/generator.h"
 #include "arrow/testing/gtest_util.h"
+#include "arrow/testing/matchers.h"
 #include "arrow/testing/util.h"
 #include "arrow/util/range.h"
 #include "arrow/util/vector.h"
