+++ cpp/src/arrow/result_test.cc
@@ -27,6 +27,7 @@
 
 #include "arrow/testing/gtest_compat.h"
 #include "arrow/testing/gtest_util.h"
+#include "arrow/testing/matchers.h"
 
 namespace arrow {
